refactored naavigation folder structure
